requirements.txt
ultralytics==8.0.20
torch>=1.9.0
torchvision>=0.10.0
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
imbalanced-learn
albumentations
tqdm
psutil
Pillow